requirements.txt
fastapi
uvicorn
deepface
pillow
tf-keras
python-multipart
python-dotenv
qdrant-client
ultralytics